Fixing detecting v1/v2 backend for v2 only policy
When calling GET backend_path/... with a v1 style url while a v2
backend is configured with according policies (containing /data in the
path) the URL failed (err != nil) and returns a string explaining 403
blablabla.
In that case, we need to check with a V2 style url before returning an
error... Otherwise, it's impossible to create a secret from a v2 backend
because of the 403 returned.